feat(LoginService.cs): adding isLoginCorrect method
